misc.c: Address warning -Wsign-compare
The symptom:
> src/misc.c: In function ‘path_starts_with’:
> src/misc.c:169:45: error: operand of ‘?:’ changes signedness from ‘long int’ to ‘long unsigned int’ due to unsignedness of other operand [-Werror=sign-compare]
>   169 |         size_t path_part_len = path_slash ? path_slash - path_part : path_len - (path_part - path);
>       |                                             ^~~~~~~~~~~~~~~~~~~~~~

+++ src/misc.c
@@ -166,7 +166,7 @@ bool path_starts_with(const char *path, const char* prefix, size_t prefix_len)
 
         const char* path_part = path + (prefix_part - prefix);
         const char* path_slash = strchr(path_part, '/');
-        size_t path_part_len = path_slash ? path_slash - path_part : path_len - (path_part - path);
+        size_t path_part_len = path_slash ? (size_t)(path_slash - path_part) : path_len - (path_part - path);
 
         return prefix_part_len == path_part_len;
     }
